requirements.txt
Flask
torch
torchvision
opencv-python
Pillow
numpy
ultralytics
gunicorn
gitpython>=3.1.30
setuptools>=70.0.0
flask-cors
python-dotenv
google.generativeai
